(feature) implement JPA Repository for Category class
