--- src/main/java/pl/glozaaleksandra/homebudget/Person.java
@@ -1,16 +1,0 @@
-package pl.glozaaleksandra.homebudget;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@AllArgsConstructor
-@Getter
-@EqualsAndHashCode
-@ToString
-@Builder
-public class Person {
-    private String name;
-}
