--- src/main/java/pl/glozaaleksandra/homebudget/Statistic.java
@@ -1,4 +1,0 @@
-package pl.glozaaleksandra.homebudget;
-
-public class Statistic {
-}
