+++ src/main/java/pl/glozaaleksandra/homebudget/controllers/category/CategoryControllery.java
@@ -3,7 +3,7 @@ package pl.glozaaleksandra.homebudget.controllers.category;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.glozaaleksandra.homebudget.repository.category.Category;
+import pl.glozaaleksandra.homebudget.model.Category;
 import pl.glozaaleksandra.homebudget.service.category.CategoryService;
 
 import java.util.List;
