+++ src/main/java/pl/glozaaleksandra/homebudget/entities/CategoryEntity.java
@@ -1,0 +1,26 @@
+package pl.glozaaleksandra.homebudget.entities;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Table(name = "category")
+@Entity
+public class CategoryEntity {
+    @Id
+    @Column(name = "category_id")
+    private int categoryId;
+
+    @Column(name = "category_name")
+    private String categoryName;
+}
