+++ src/main/java/pl/glozaaleksandra/homebudget/model/Category.java
@@ -1,0 +1,13 @@
+package pl.glozaaleksandra.homebudget.model;
+
+import lombok.*;
+
+@ToString
+@Getter
+@Setter
+@EqualsAndHashCode
+@Builder
+@AllArgsConstructor
+public class Category {
+    private String name;
+}
