+++ src/main/java/pl/glozaaleksandra/homebudget/model/Person.java
@@ -1,0 +1,12 @@
+package pl.glozaaleksandra.homebudget.model;
+
+import lombok.*;
+
+@AllArgsConstructor
+@Getter
+@EqualsAndHashCode
+@ToString
+@Builder
+public class Person {
+    private String name;
+}
