+++ src/main/java/pl/glozaaleksandra/homebudget/model/Statistic.java
@@ -1,0 +1,4 @@
+package pl.glozaaleksandra.homebudget.model;
+
+public class Statistic {
+}
