+++ src/main/java/pl/glozaaleksandra/homebudget/nodatabase/repository/category/factory/CategoryRepositoryFactory.java
@@ -1,0 +1,7 @@
+package pl.glozaaleksandra.homebudget.nodatabase.repository.category.factory;
+
+import pl.glozaaleksandra.homebudget.nodatabase.repository.category.CategoryRepository;
+
+public interface CategoryRepositoryFactory {
+  public CategoryRepository createCategoryRepository();
+}
