+++ src/main/java/pl/glozaaleksandra/homebudget/nodatabase/repository/expense/ListBasedExpenseRepository.java
@@ -1,6 +1,7 @@
-package pl.glozaaleksandra.homebudget.repository.expense;
+package pl.glozaaleksandra.homebudget.nodatabase.repository.expense;
 
-import pl.glozaaleksandra.homebudget.Person;
+import pl.glozaaleksandra.homebudget.model.Expense;
+import pl.glozaaleksandra.homebudget.model.Person;
 
 import java.util.ArrayList;
 import java.util.List;
