+++ src/main/java/pl/glozaaleksandra/homebudget/nodatabase/repository/person/PersonRepository.java
@@ -1,12 +1,15 @@
-package pl.glozaaleksandra.homebudget.repository.person;
+package pl.glozaaleksandra.homebudget.nodatabase.repository.person;
+
+import pl.glozaaleksandra.homebudget.model.Person;
 
 import java.util.List;
 import java.util.Optional;
-import pl.glozaaleksandra.homebudget.Person;
 
 public interface PersonRepository {
 
   Person addNewPerson(String name);
+
   List<Person> findAllPersons();
+
   Optional<Person> getByName(String name);
 }
