+++ src/main/java/pl/glozaaleksandra/homebudget/nodatabase/repository/product/ProductRepository.java
@@ -1,4 +1,6 @@
-package pl.glozaaleksandra.homebudget.repository.product;
+package pl.glozaaleksandra.homebudget.nodatabase.repository.product;
+
+import pl.glozaaleksandra.homebudget.model.Product;
 
 import java.util.List;
 import java.util.Optional;
