+++ src/main/java/pl/glozaaleksandra/homebudget/repository/CategoryRepository.java
@@ -1,0 +1,10 @@
+package pl.glozaaleksandra.homebudget.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import pl.glozaaleksandra.homebudget.entities.CategoryEntity;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
+
+}
