--- src/main/java/pl/glozaaleksandra/homebudget/repository/category/Category.java
@@ -1,19 +1,0 @@
-package pl.glozaaleksandra.homebudget.repository.category;
-
-import java.util.Objects;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@ToString
-@Getter
-@Setter
-@EqualsAndHashCode
-@Builder
-@AllArgsConstructor
-public class Category {
-    private String name;
-}
