--- src/main/java/pl/glozaaleksandra/homebudget/repository/category/factory/CategoryRepositoryFactory.java
@@ -1,7 +1,0 @@
-package pl.glozaaleksandra.homebudget.repository.category.factory;
-
-import pl.glozaaleksandra.homebudget.repository.category.CategoryRepository;
-
-public interface CategoryRepositoryFactory {
-  public CategoryRepository createCategoryRepository();
-}
