--- src/main/java/pl/glozaaleksandra/homebudget/repository/expense/ExpenseService.java
@@ -1,8 +1,0 @@
-package pl.glozaaleksandra.homebudget.repository.expense;
-
-import pl.glozaaleksandra.homebudget.repository.category.CategoryRepository;
-import pl.glozaaleksandra.homebudget.service.category.CategoryService;
-
-public class ExpenseService {
-  private CategoryService categoryService;
-}
