+++ src/main/resources/application.properties
@@ -1,1 +1,4 @@
 spring.application.name=homebudget
+spring.datasource.url=jdbc:postgresql://localhost:5555/homebudget
+spring.datasource.username=user
+spring.datasource.password=test
